module_summarization/pyproject.toml
[tool.poetry]
name = "lca"
version = "0.1.0"
description = "[WIP] Benchmarking code models"
authors = [
    "Egor Bogomolov <egor.bogomolov@jetbrains.com>",
    "Maria Tigina <maria.tigina@jetbrains.com>"
]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/JetBrains-Research/lca"
package-mode = false

[tool.poetry.dependencies]
python = "~3.10"
black = {extras = ["jupyter"], version = "^23.7.0"}
isort = "^5.12.0"
mypy = "^1.5.0"
jupyter = "^1.0.0"
pandas = "^2.0.3"
matplotlib = "^3.7.2"
seaborn = "^0.12.2"
aiohttp = "^3.8.5"
tenacity = "^8.2.3"
requests = "^2.31.0"
lxml = "^4.9.3"
lxml-stubs = "^0.4.0"
self = "^2020.12.3"
pydriller = "^2.5"
pyyaml = "^6.0.1"
nltk = "^3.8.1"
gensim = "^4.3.1"
scikit-learn = "^1.3.0"
pandas-stubs = "^2.0.3.230814"
types-pyyaml = "^6.0.12.11"
openai = "^2.8.1"
transformers = "^4.33.1"
torch = "2.0"
evaluate = "^0.4.0"
sacrebleu = "^2.3.1"
absl-py = "^2.0.0"
rouge-score = "^0.1.2"
bert-score = "^0.3.13"
docutils = "0.17"
sphinx = "4.5.0"
rst2txt = "^1.1.0"
pandoc = "^2.3"
accelerate = "^0.25.0"
bitsandbytes = "^0.41.2.post2"
llama-index = "^0.9.36"
openpyxl = "^3.1.2"
together = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120
target-version = ["py310"]

[tool.isort]
line_length = 120
py_version = 310
profile = "black"

[tool.mypy]
python_version = "3.10"

[[tool.mypy.overrides]]
module = []
ignore_missing_imports = true
